requirements.txt
python-telegram-bot
mutagen
pydub
schedule